requirements.txt
# AI_GENERATED_CODE_START
# [AI Generated] Data: 19/03/2024
# Descrição: Arquivo de dependências para projeto LangChain com OpenAI
# Gerado por: Cursor AI
# Versão: Python 3.10+, langchain 0.1.0+, openai 1.13.3+

langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai>=1.13.3
python-dotenv==1.0.0
wikipedia==1.4.0
numexpr==2.8.7 